Avoid different EVC instances in schedule and memory
When an EVC was loaded from storehouse on startup, a test checked
if that EVC already existed in the circuits dictionary in main.py.
But, as load_evcs can be called multiple times concurrently, that
check alone could not avoid duplicates being created.
To solve this, I used the setdefault method of the dictionary to
set the EVC to the dictionary. It only sets the value if it
does not already exists. Only if the EVC was set then the
scheduler is add. Fixes #166.